requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
Werkzeug==3.0.3
flask-cors
selenium
BeautifulSoup4
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.30
psycopg2==2.9.9
psycopg2-binary==2.9.9
requests==2.32.3
requests-cache==0.5.2
python-dotenv
gunicorn